requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
idna==3.4
mysqlclient==2.1.1
pycodestyle==2.9.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
